+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## v0.0.3
+
+新增命令：`newThemeEditProcessorHandler`: 自动将新创建的Processor里的Handler函数的方法签名修改为点引用，并在参数列表中加上开头的“board”，避免歧义。
+
+fileUtils新增将某文件指定行替换为给定字符串的工具函数；commonUtils新增给定主题ID，获取ControllerName和ThemeName的工具函数。
+（剩余功能TODO...）s
+
 ## v0.0.2
 
 封装输出提示信息的API，不仅弹出通知提示，还同时输出到output.